Refactor to ecommerce package
