+++ jobs/inference/run_reviews_preprocess.py
@@ -1,5 +1,5 @@
-from utils import *
-from preprocess.reviews import *
+from ecommerce.utils import *
+from ecommerce.preprocess.reviews import *
 
 if __name__ == "__main__":
     dataset, path = get_dataset(BronzeDataName.ORDER_REVIEWS)
