+++ pyproject.toml
@@ -29,7 +29,7 @@ build-backend = "poetry.core.masonry.api"
 
 [tool.poetry]
 packages = [
-  { include = "*", from = "src" }
+  { include = "ecommerce", from = "src" }
 ]
 
 [tool.poetry.group.dev.dependencies]
