--- src/__init__.py
@@ -1,5 +1,0 @@
-from .utils import *
-from .features import *
-from .inference import *
-from .preprocess import *
-from .postprocess import *
