+++ src/ecommerce/inference/base.py
