+++ src/ecommerce/utils/paths.py
