+++ tests/common/test_config.py
@@ -6,7 +6,7 @@ from pathlib import Path
 import yaml
 import shutil
 
-from utils.config import PreprocessConfig, TranslatePipelineConfig
+from ecommerce.utils.config import PreprocessConfig, TranslatePipelineConfig
 
 ARTIFACT_PATH = Path("./tests/tmp_config_test/")
 ARTIFACT_PATH.mkdir(parents=True, exist_ok=True)
